requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.testing.in, then run this to update:
#
#    pip-compile requirements.testing.in
#
astroid==1.4.6            # via pylint
click==6.6
colorama==0.3.7           # via pylint
cov-core==1.13.0          # via pytest-cov
coverage==3.7.1           # via cov-core
lazy-object-proxy==1.2.2  # via astroid
mock==1.0.1
psycopg2==2.6.1
py==1.4.30                # via pytest
pylint==1.5.4
pytest-cov==1.7.0
pytest==2.8.7
pyyaml==3.11
requests==2.10.0
six==1.10.0               # via astroid, pylint
wrapt==1.10.8             # via astroid
